+++ CHANGELOG.md
@@ -60,6 +60,10 @@ Released initial draft of *json-xform*
 
 ----
 
+***missing notes***
+
+----
+
 ## 1.5.1
 ### Oct 9
 #### Update Dependencies; Accept `--extends` command-line option